+++ src/applications/merchant-app/MerchantPopulateItemsTab.svelte
@@ -6,7 +6,6 @@
   import CustomDialog from "../components/CustomDialog.svelte";
   import ItemEntry from "./ItemEntry.svelte";
   import * as PileUtilities from "../../helpers/pile-utilities.js";
-  import * as Utilities from "../../helpers/utilities.js";
   import { getSetting } from "../../helpers/helpers";
   import { slide } from "svelte/transition";
   import { quintOut } from "svelte/easing";
@@ -132,7 +131,6 @@
         const existingItem = items.find(
           (item) => item.documentId === newItem.documentId
         );
-        console.log(items, existingItem?.quantity, newItem.quantity);
         if (existingItem) {
           existingItem.quantity += newItem.quantity;
         } else {